refactor(llm): provider_client 拆分为多子模块包
将 1451 行 provider_client.py 拆为 errors/urls/auth/budget/mapping/models/client 子模块，经 __init__ 全量重导出保持导入面。子模块经包命名空间在调用时解析共享符号，保证测试打桩传播。44 相关测试通过。
